Recursively enumerate rule results for matches
Currently Heket isn't emitting the nested AST for rule results, so it's hard
to use the result object to generate an opinionated AST that matches the input
because we can only flesh out that AST structure a few levels deep. After that
the nested data isn't there.

This fixes the problem but only works when the regex caching feature is diabled.